feat(nurse): add useHospitalAdmissions shared hook
Extracts the GET /inpatient/admissions?hospitalId= fetch + mapping
logic out of nurse/patients into a reusable hook in src/lib/hospital.ts,
so nurse/notes and nurse/dashboard can consume the same live patient
list instead of copy-pasting the fetch. patients/page.tsx refactored
onto it with no behavior change (same dev-only mock fallback).
